notes on disabled vertex cache and async shader compiler
split log result formatter in multiple files for easier navigation
