+++ CompatBot/Utils/ResultFormatters/LogParserResultFormatter.cs
@@ -19,7 +19,7 @@ using IrdLibraryClient.IrdFormat;
 
 namespace CompatBot.Utils.ResultFormatters
 {
-    internal static class LogParserResult
+    internal static partial class LogParserResult
     {
         private static readonly Client compatClient = new Client();
         private static readonly IrdClient irdClient = new IrdClient();
@@ -54,6 +54,8 @@ namespace CompatBot.Utils.ResultFormatters
             //{"BCJS30022", "NPJA00102"},
             {"BCJS70013", "NPJA00102"},
         };
+
+        private static readonly string[] KnownDisableVertexCacheIds = { "NPEB00258", "NPUB30162", "NPJB00068" };
 
         private static readonly HashSet<string> KnownBogusLicenses = new HashSet<string>(StringComparer.InvariantCultureIgnoreCase)
         {
@@ -207,456 +209,6 @@ namespace CompatBot.Utils.ResultFormatters
                     value = DisabledMark;
                 items[key] = value.Sanitize(false);
             }
-        }
-
-        private static void BuildInfoSection(DiscordEmbedBuilder builder, NameValueCollection items)
-        {
-            var systemInfo = items["build_and_specs"] ?? "";
-            var m = BuildInfoInLog.Match(systemInfo);
-            if (m.Success)
-            {
-                items["build_branch"] = m.Groups["branch"].Value.Trim();
-                items["build_commit"] = m.Groups["commit"].Value.Trim();
-                items["fw_version_installed"] = m.Groups["fw_version_installed"].Value;
-                items["cpu_model"] = m.Groups["cpu_model"].Value
-                    .Replace("(R)", "", StringComparison.InvariantCultureIgnoreCase)
-                    .Replace("®", "", StringComparison.InvariantCultureIgnoreCase)
-                    .Replace("(TM)", "", StringComparison.InvariantCultureIgnoreCase)
-                    .Replace("™", "", StringComparison.InvariantCultureIgnoreCase)
-                    .Replace(" CPU", "");
-                items["thread_count"] = m.Groups["thread_count"].Value;
-                items["memory_amount"] = m.Groups["memory_amount"].Value;
-                items["cpu_extensions"] = m.Groups["cpu_extensions"].Value;
-                systemInfo = $"RPCS3 v{m.Groups["version_string"].Value} {m.Groups["stage"].Value} | {m.Groups["branch"].Value}";
-                if (!string.IsNullOrEmpty(items["fw_version_installed"]))
-                    systemInfo += " | FW " + items["fw_version_installed"];
-                if (!string.IsNullOrEmpty(items["os_path"]))
-                    systemInfo += " | " + items["os_path"];
-                systemInfo += $"{Environment.NewLine}{items["cpu_model"]} | {items["thread_count"]} Threads | {items["memory_amount"]} GiB RAM";
-                if (!string.IsNullOrEmpty(items["cpu_extensions"]))
-                    systemInfo += " | " + items["cpu_extensions"];
-            }
-            if (items["gpu_info"] is string gpu)
-                systemInfo += $"{Environment.NewLine}GPU: {gpu}";
-            else if (items["gpu_available_info"] is string availableGpus)
-            {
-                var multiple = availableGpus.Contains(Environment.NewLine);
-                systemInfo +=$"{Environment.NewLine}GPU{(multiple ? "s" : "")}:{(multiple ? Environment.NewLine : " ")}{availableGpus}";
-            }
-            builder.AddField("Build Info", systemInfo.Trim(EmbedPager.MaxFieldLength));
-        }
-
-        private static (string name, List<string> lines) BuildCpuSection(NameValueCollection items)
-        {
-            if (string.IsNullOrEmpty(items["ppu_decoder"]))
-                return (null, null);
-
-            var lines = new List<string>
-            {
-                $"PPU Decoder: {items["ppu_decoder"],21}",
-                $"SPU Decoder: {items["spu_decoder"],21}",
-                $"SPU Lower Thread Priority: {items["spu_lower_thread_priority"],7}",
-                $"SPU Loop Detection: {items["spu_loop_detection"],14}",
-                $"Thread Scheduler: {items["thread_scheduler"],16}",
-                $"SPU Threads: {items["spu_threads"],21}",
-                $"SPU Block Size: {items["spu_block_size"] ?? "N/A",18}",
-                $"Accurate xfloat: {items["accurate_xfloat"] ?? "N/A",17}",
-                $"Force CPU Blit: {items["cpu_blit"] ?? "N/A",18}",
-                $"Lib Loader: {items["lib_loader"],22}",
-            };
-            return ("CPU Settings", lines);
-        }
-
-        private static (string name, List<string> lines) BuildGpuSection(NameValueCollection items)
-        {
-            if (string.IsNullOrEmpty(items["renderer"]))
-                return (null, null);
-
-            var lines = new List<string>
-            {
-
-                $"Renderer: {items["renderer"], 24}",
-                $"Aspect ratio: {items["aspect_ratio"], 20}",
-                $"Resolution: {items["resolution"], 22}",
-                $"Resolution Scale: {items["resolution_scale"] ?? "N/A", 16}",
-                $"Resolution Scale Threshold: {items["texture_scale_threshold"] ?? "N/A", 6}",
-                $"Write Color Buffers: {items["write_color_buffers"], 13}",
-                $"Anisotropic Filter: {items["af_override"] ?? "N/A", 14}",
-                $"Frame Limit: {items["frame_limit"], 21}",
-                $"Disable Async Shaders: {items["async_shaders"] ?? "N/A", 11}",
-                $"Disable Vertex Cache: {items["vertex_cache"], 12}",
-            };
-            return ("GPU Settings", lines);
-        }
-
-        private static void BuildSettingsSections(DiscordEmbedBuilder builder, NameValueCollection items, (string name, List<string> lines) colA, (string name, List<string> lines) colB)
-        {
-            if (colA.lines?.Count > 0 && colB.lines?.Count > 0)
-            {
-                var isCustomSettings = items["custom_config"] != null;
-                var colAToRemove = colA.lines.Count(l => l.EndsWith("N/A"));
-                var colBToRemove = colB.lines.Count(l => l.EndsWith("N/A"));
-                var linesToRemove = Math.Min(colAToRemove, colBToRemove);
-                if (linesToRemove > 0)
-                {
-                    var linesToSkip = colAToRemove - linesToRemove;
-                    var tmp = colA.lines;
-                    colA.lines = new List<string>(tmp.Count - linesToRemove);
-                    for (var i = 0; i < tmp.Count; i++)
-                        if (!tmp[i].EndsWith("N/A") || (linesToSkip--) > 0)
-                            colA.lines.Add(tmp[i]);
-
-                    linesToSkip = colBToRemove - linesToRemove;
-                    tmp = colB.lines;
-                    colB.lines = new List<string>(tmp.Count - linesToRemove);
-                    for (var i = 0; i < tmp.Count; i++)
-                        if (!tmp[i].EndsWith("N/A") || (linesToSkip--) > 0)
-                            colB.lines.Add(tmp[i]);
-                }
-                AddSettingsSection(builder, colA.name, colA.lines, isCustomSettings);
-                AddSettingsSection(builder, colB.name, colB.lines, isCustomSettings);
-            }
-        }
-
-        private static void AddSettingsSection(DiscordEmbedBuilder builder, string name, List<string> lines, bool isCustomSettings)
-        {
-            var result = new StringBuilder();
-            foreach (var line in lines)
-                result.Append("`").Append(line).AppendLine("`");
-            if (isCustomSettings)
-                name = "Per-game " + name;
-            builder.AddField(name, result.ToString().FixSpaces(), true);
-        }
-
-        private static void BuildLibsSection(DiscordEmbedBuilder builder, NameValueCollection items)
-        {
-            if (items["lib_loader"] is string libs && libs.Contains("manual", StringComparison.InvariantCultureIgnoreCase))
-                builder.AddField("Selected Libraries", items["library_list"]?.Trim(1024));
-        }
-
-        private static void BuildWeirdSettingsSection(DiscordEmbedBuilder builder, NameValueCollection items)
-        {
-            var notes = new List<string>();
-            if (!string.IsNullOrWhiteSpace(items["log_disabled_channels"]))
-                notes.Add("❗ Some logging priorities were modified, please reset and upload a new log");
-            if (!string.IsNullOrEmpty(items["resolution"]) && items["resolution"] != "1280x720")
-                notes.Add("⚠ `Resolution` was changed from the recommended `1280x720`");
-            if (items["hook_static_functions"] is string hookStaticFunctions && hookStaticFunctions == EnabledMark)
-                notes.Add("⚠ `Hook Static Functions` is enabled, please disable");
-            if (items["host_root"] is string hostRoot && hostRoot == EnabledMark)
-                notes.Add("❔ `/host_root/` is enabled");
-            if (items["gpu_texture_scaling"] is string gpuTextureScaling && gpuTextureScaling == EnabledMark)
-                notes.Add("⚠ `GPU Texture Scaling` is enabled, please disable");
-            if (items["af_override"] is string af)
-            {
-                if (af == "Disabled")
-                    notes.Add("❌ `Anisotropic Filter` is `Disabled`, please use `Auto` instead");
-                else if (af.ToLowerInvariant() != "auto" && af != "16")
-                    notes.Add($"❔ `Anisotropic Filter` is set to `{af}x`, which makes little sense over `16x` or `Auto`");
-            }
-            if (items["resolution_scale"] is string resScale && int.TryParse(resScale, out var resScaleFactor) && resScaleFactor < 100)
-                notes.Add($"❔ `Resolution Scale` is `{resScale}%`; this will not increase performance");
-            if (items["cpu_blit"] is string cpuBlit && cpuBlit == EnabledMark && items["write_color_buffers"] is string wcb && wcb == DisabledMark)
-                notes.Add("⚠ `Force CPU Blit` is enabled, but `Write Color Buffers` is disabled");
-            if (items["zcull"] is string zcull && zcull == EnabledMark)
-                notes.Add("⚠ `ZCull Occlusion Queries` are disabled, can result in visual artifacts");
-            if (items["driver_recovery_timeout"] is string driverRecoveryTimeout && int.TryParse(driverRecoveryTimeout, out var drtValue) && drtValue != 1000000)
-            {
-                if (drtValue == 0)
-                    notes.Add("⚠ `Driver Recovery Timeout` is set to 0 (infinite), please use default value of 1000000");
-                else if (drtValue < 10_000)
-                    notes.Add($"⚠ `Driver Recovery Timeout` is set too low: {GetTimeFormat(drtValue)} (1 frame @ {(1_000_000.0 / drtValue):0.##} fps)");
-                else if (drtValue > 10_000_000)
-                    notes.Add($"⚠ `Driver Recovery Timeout` is set too high: {GetTimeFormat(drtValue)}");
-            }
-            if (items["hle_lwmutex"] is string hleLwmutex && hleLwmutex == EnabledMark)
-                notes.Add("⚠ `HLE lwmutex` is enabled, might affect compatibility");
-            if (items["spu_block_size"] is string spuBlockSize)
-            {
-/*
-                if (spuBlockSize == "Giga")
-                    notes.AppendLine("`Giga` mode for `SPU Block Size` is strongly not recommended to use");
-*/
-                if (spuBlockSize != "Safe")
-                    notes.Add($"⚠ Please use `Safe` mode for `SPU Block Size`. `{spuBlockSize}` is currently unstable.");
-            }
-
-            if (items["lib_loader"] is string libLoader
-                && libLoader.Contains("Auto", StringComparison.InvariantCultureIgnoreCase)
-                && (libLoader == "Auto"
-                    || (libLoader.Contains("manual", StringComparison.InvariantCultureIgnoreCase) && string.IsNullOrEmpty(items["library_list"]))))
-            {
-                notes.Add("⚠ Please use `Load liblv2.sprx only` as a `Library loader`");
-            }
-
-            var notesContent = new StringBuilder();
-            foreach (var line in SortLines(notes))
-                notesContent.AppendLine(line);
-            PageSection(builder, notesContent.ToString().Trim(), "Important Settings to Review");
-        }
-
-        private static void BuildMissingLicensesSection(DiscordEmbedBuilder builder, NameValueCollection items)
-        {
-            if (items["rap_file"] is string rap)
-            {
-                var limitTo = 5;
-                var licenseNames = rap.Split(Environment.NewLine)
-                    .Distinct()
-                    .Select(Path.GetFileName)
-                    .Distinct()
-                    .Except(KnownBogusLicenses)
-                    .Select(p => $"`{p}`")
-                    .ToList();
-                if (licenseNames.Count == 0)
-                    return;
-
-                string content;
-                if (licenseNames.Count > limitTo)
-                {
-                    content = string.Join(Environment.NewLine, licenseNames.Take(limitTo - 1));
-                    var other = licenseNames.Count - limitTo + 1;
-                    content += $"{Environment.NewLine}and {other} other license{StringUtils.GetSuffix(other)}";
-                }
-                else
-                    content = string.Join(Environment.NewLine, licenseNames);
-                builder.AddField("Missing Licenses", content);
-            }
-        }
-
-        private static async Task<(bool irdChecked, bool broken)> HasBrokenFilesAsync(NameValueCollection items)
-        {
-            if (!(items["serial"] is string productCode))
-                return (false, false);
-
-            if (!productCode.StartsWith("B") && !productCode.StartsWith("M"))
-                return (false, false);
-
-            if (string.IsNullOrEmpty(items["broken_directory"])
-                && string.IsNullOrEmpty(items["broken_filename"]))
-                return (false, false);
-
-            var getIrdTask = irdClient.DownloadAsync(productCode, Config.IrdCachePath, Config.Cts.Token);
-            var missingDirs = items["broken_directory"]?.Split(Environment.NewLine).Distinct().ToList() ?? new List<string>(0);
-            var missingFiles = items["broken_filename"]?.Split(Environment.NewLine).Distinct().ToList() ?? new List<string>(0);
-            HashSet<string> knownFiles;
-            try
-            {
-                var irdFiles = await getIrdTask.ConfigureAwait(false);
-                knownFiles = new HashSet<string>(
-                    from ird in irdFiles
-                    from name in ird.GetFilenames()
-                    select name,
-                    StringComparer.InvariantCultureIgnoreCase
-                );
-            }
-            catch (Exception e)
-            {
-                Config.Log.Warn(e, "Failed to get IRD files for " + productCode);
-                return (false, false);
-            }
-            if (knownFiles.Count == 0)
-                return (false, false);
-
-            var broken = missingFiles.Any(knownFiles.Contains);
-            if (broken)
-                return (true, true);
-
-            var knownDirs = new HashSet<string>(knownFiles.Select(f => Path.GetDirectoryName(f).Replace('\\', '/')), StringComparer.InvariantCultureIgnoreCase);
-            return (true, missingDirs.Any(knownDirs.Contains));
-        }
-
-        private static async Task BuildNotesSectionAsync(DiscordEmbedBuilder builder, LogParseState state, NameValueCollection items, DiscordClient discordClient)
-        {
-            BuildWeirdSettingsSection(builder, items);
-            BuildMissingLicensesSection(builder, items);
-            var (irdChecked, brokenDump) = await HasBrokenFilesAsync(items).ConfigureAwait(false);
-            brokenDump |= !string.IsNullOrEmpty(items["edat_block_offset"]);
-            var elfBootPath = items["elf_boot_path"] ?? "";
-            var isEboot = !string.IsNullOrEmpty(elfBootPath) && elfBootPath.EndsWith("EBOOT.BIN", StringComparison.InvariantCultureIgnoreCase);
-            var isElf = !string.IsNullOrEmpty(elfBootPath) && !elfBootPath.EndsWith("EBOOT.BIN", StringComparison.InvariantCultureIgnoreCase);
-            var notes = new List<string>();
-            if (items["fatal_error"] is string fatalError)
-            {
-                builder.AddField("Fatal Error", $"```{fatalError.Trim(1022)}```");
-                if (fatalError.Contains("psf.cpp") || fatalError.Contains("invalid map<K, T>"))
-                    notes.Add("⚠ Game save data might be corrupted");
-                else if (fatalError.Contains("Could not bind OpenGL context"))
-                    notes.Add("❌ GPU or installed GPU drivers do not support OpenGL 4.3");
-            }
-            if (items["failed_to_decrypt"] is string _)
-                notes.Add("❌ Failed to decrypt game content, license file might be corrupted");
-            if (items["failed_to_boot"] is string _)
-                notes.Add("❌ Failed to boot the game, the dump might be encrypted or corrupted");
-            if (brokenDump)
-                notes.Add("❌ Some game files are missing or corrupted, please re-dump and validate.");
-            else if (irdChecked)
-                notes.Add("✅ Checked missing files against IRD");
-            if (items["fw_version_installed"] is string fw && !string.IsNullOrEmpty(fw))
-            {
-                if (Version.TryParse(fw, out var fwv))
-                {
-                    if (fwv < MinimumFirmwareVersion)
-                        notes.Add($"⚠ Firmware version {MinimumFirmwareVersion} or later is recommended");
-                }
-                else
-                    notes.Add("⚠ Custom firmware is not supported, please use the latest official one");
-            }
-
-            if (!string.IsNullOrEmpty(items["host_root_in_boot"]) && isEboot)
-                notes.Add("❌ Retail game booted as an ELF through the `/root_host/`, probably due to passing path as an argument; please boot through the game library list for now");
-            if (!string.IsNullOrEmpty(items["serial"]) && isElf)
-                notes.Add($"⚠ Retail game booted directly through `{Path.GetFileName(elfBootPath)}`, which is not recommended");
-            if (string.IsNullOrEmpty(items["serial"] + items["game_title"]) && items["fw_version_installed"] is string fwVersion)
-            {
-                notes.Add($"ℹ The log contains only installation of firmware {fwVersion}");
-                notes.Add("ℹ Please boot the game and upload a new log");
-            }
-            if (string.IsNullOrEmpty(items["ppu_decoder"]) || string.IsNullOrEmpty(items["renderer"]))
-            {
-                notes.Add("ℹ The log is empty");
-                notes.Add("ℹ Please boot the game and upload a new log");
-            }
-
-            if (items["cpu_model"] is string cpu)
-            {
-                if (cpu.StartsWith("AMD"))
-                {
-                    if (!cpu.Contains("Ryzen"))
-                        notes.Add("⚠ AMD CPUs before Ryzen are too weak for PS3 emulation");
-                }
-                if (cpu.StartsWith("Intel"))
-                {
-                    if (cpu.Contains("Core2") || cpu.Contains("Celeron") || cpu.Contains("Atom"))
-                        notes.Add("⚠ This CPU is too old and/or too weak for PS3 emulation");
-                }
-            }
-            if (int.TryParse(items["thread_count"], out var threadCount) && threadCount < 4)
-                notes.Add($"⚠ This CPU only has {threadCount} hardware thread{(threadCount == 1 ? "" : "s")} enabled");
-
-            var supportedGpu = true;
-            Version oglVersion = null;
-            if (items["opengl_version"] is string oglVersionString)
-                Version.TryParse(oglVersionString, out oglVersion);
-            if (items["glsl_version"] is string glslVersionString && Version.TryParse(glslVersionString, out var glslVersion))
-            {
-                glslVersion = new Version(glslVersion.Major, glslVersion.Minor/10);
-                if (oglVersion == null || glslVersion > oglVersion)
-                    oglVersion = glslVersion;
-            }
-            if (oglVersion != null)
-            {
-                if (oglVersion < MinimumOpenGLVersion)
-                {
-                    notes.Add($"❌ GPU only supports OpenGL {oglVersion.Major}.{oglVersion.Minor}, which is below the minimum requirement of {MinimumOpenGLVersion}");
-                    supportedGpu = false;
-                }
-            }
-            if (supportedGpu
-                && items["gpu_info"] is string gpuInfo)
-            {
-                if (IntelGpuModel.Match(gpuInfo) is Match intelMatch
-                    && intelMatch.Success)
-                {
-                    var modelNumber = intelMatch.Groups["gpu_model_number"].Value;
-                    if (!string.IsNullOrEmpty(modelNumber) && modelNumber.StartsWith('P'))
-                        modelNumber = modelNumber.Substring(1);
-                    int.TryParse(modelNumber, out var modelNumberInt);
-                    if (modelNumberInt < 500 || modelNumberInt > 1000)
-                    {
-                        notes.Add("❌ Intel iGPUs before Skylake do not fully comply with OpenGL 4.3");
-                        supportedGpu = false;
-                    }
-                    else
-                        notes.Add("⚠ Intel iGPUs are not officially supported, visual glitches are to be expected");
-                }
-                if (items["os_path"] is string os
-                    && os != "Linux"
-                    && IsNvidia(gpuInfo)
-                    && items["driver_version_info"] is string driverVersionString
-                    && Version.TryParse(driverVersionString, out var driverVersion))
-                {
-                    if (driverVersion < NvidiaRecommendedOldWindowsVersion)
-                        notes.Add($"⚠ Please update your nVidia driver to at least {NvidiaRecommendedOldWindowsVersion}");
-                    if (driverVersion >= NvidiaFullscreenBugMinVersion
-                        && driverVersion < NvidiaFullscreenBugMaxVersion
-                        && items["renderer"] == "Vulkan"
-                        && items["vsync"] == EnabledMark)
-                        notes.Add("⚠ **400 series** nVidia drivers can cause random screen freeze when playing in **fullscreen** using **Vulkan** renderer with **vsync disabled**");
-                }
-            }
-            if (!string.IsNullOrEmpty(items["shader_compile_error"]))
-            {
-                if (supportedGpu)
-                    notes.Add("❌ Shader compilation error might indicate shader cache corruption");
-                else
-                    notes.Add("❌ Shader compilation error on unsupported GPU");
-            }
-
-            if (!string.IsNullOrEmpty(items["ppu_hash_patch"]) || !string.IsNullOrEmpty(items["spu_hash_patch"]))
-                notes.Add("ℹ Game-specific patches were applied");
-
-            bool discInsideGame = false;
-            bool discAsPkg = false;
-            if (items["game_category"] == "DG")
-            {
-                discInsideGame |= !string.IsNullOrEmpty(items["ldr_disc"]) && !(items["serial"]?.StartsWith("NP", StringComparison.InvariantCultureIgnoreCase) ?? false);
-                discAsPkg |= items["serial"]?.StartsWith("NP", StringComparison.InvariantCultureIgnoreCase) ?? false;
-                discAsPkg |= items["ldr_game_serial"] is string ldrGameSerial
-                             && ldrGameSerial.StartsWith("NP", StringComparison.InvariantCultureIgnoreCase);
-            }
-            discAsPkg |= items["game_category"] == "HG" && !(items["serial"]?.StartsWith("NP", StringComparison.InvariantCultureIgnoreCase) ?? false);
-            if (discInsideGame)
-                notes.Add($"❌ Disc game inside `{items["ldr_disc"]}`");
-            DiscordEmoji pirateEmoji = null;
-            if (discAsPkg)
-            {
-                pirateEmoji = discordClient.GetEmoji(":piratethink:", DiscordEmoji.FromUnicode("🔨"));
-                notes.Add($"{pirateEmoji} Disc game installed as a PKG ");
-            }
-
-            if (!string.IsNullOrEmpty(items["native_ui_input"]))
-                notes.Add("⚠ Pad initialization problem detected; try disabling `Native UI`");
-            if (!string.IsNullOrEmpty(items["xaudio_init_error"]))
-                notes.Add("❌ XAudio initialization failed; make sure you have audio output device working");
-
-            if (!string.IsNullOrEmpty(items["fw_missing_msg"])
-                || !string.IsNullOrEmpty(items["fw_missing_something"]))
-                notes.Add("❌ PS3 firmware is missing or corrupted");
-
-            var updateInfo = await CheckForUpdateAsync(items).ConfigureAwait(false);
-            var buildBranch = items["build_branch"]?.ToLowerInvariant();
-            if (updateInfo != null && (buildBranch == "head" || buildBranch == "spu_perf"))
-            {
-                string prefix = "⚠";
-                string timeDeltaStr;
-                if (updateInfo.GetUpdateDelta() is TimeSpan timeDelta)
-                {
-                    timeDeltaStr = timeDelta.AsTimeDeltaDescription() + " old";
-                    if (timeDelta > PrehistoricBuild)
-                        prefix = "😱";
-                    else if (timeDelta > AncientBuild)
-                        prefix = "💢";
-                    //else if (timeDelta > VeryVeryOldBuild)
-                    //    prefix = "💢";
-                    else if (timeDelta > VeryOldBuild)
-                        prefix = "‼";
-                    else if (timeDelta > OldBuild)
-                        prefix = "❗";
-                }
-                else
-                    timeDeltaStr = "outdated";
-                notes.Add($"{prefix} This RPCS3 build is {timeDeltaStr}, please consider updating it");
-                if (buildBranch == "spu_perf")
-                    notes.Add($"ℹ `{buildBranch}` build is obsolete, current master build offers at least the same level of performance and includes many additional improvements");
-            }
-
-            if (state.Error == LogParseState.ErrorCode.SizeLimit)
-                notes.Add("ℹ The log was too large, so only the last processed run is shown");
-
-            var notesContent = new StringBuilder();
-            foreach (var line in SortLines(notes, pirateEmoji))
-                notesContent.AppendLine(line);
-            PageSection(builder, notesContent.ToString().Trim(), "Notes");
         }
 
         private static void PageSection(DiscordEmbedBuilder builder, string notesContent, string sectionName)
